app: utils: Check the length of the avars list before finalizing it
I was periodically seeing errors like

    (process:28867): GLib-CRITICAL **: g_hash_table_lookup: assertion
    `hash_table != NULL' failed

in the error log. This was due to the call to add_multipart_avar() in
set_vars() when the avars list was empty.

This would then try to add a NULL GHashTable to the list.

So only call add_multipart_avar() in set_vars() if the avars list has
at least one element in it.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>