app: utils: Reduce the amount of memory needed for mime processing
When dealing with multipart mime post data in process_mime(). We read
the post data into a memory buffer then pass it over to gmime, this then
means we have two copies of the data in memory until we finish the
multipart mime processing and free everything.

What we can do instead is free our memory buffer immediately after
passing it over to gmime, effectively halving the amount of memory
required after that point to process the mime parts.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>