src/utils.c: Undo while loop optimisation in process_mime()
As part of commit 6dc0768 ("app: Switch to the FastCGI FCGX_ API") the
do { ... } while loop condition in process_mime() was changed from
"bytes_read > 0" to "bytes_read == BUF_SIZE".

This is actually not a good paradigm to follow as there are cases where
the kernel may actually send you less data than you requested and
_still_ have more data to send. This can happen at least with pipes.

So it's best to just keep reading until we get _no_ bytes of data.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>