app: Fix potentially slow logins
When a user logs in we do a reverse DNS lookup of the users IP address
to get the hostname. We store both the IP address and hostname (if any)
of the requesting client in the utmp table. This is used in the 'Last
login' line we display. Also having both the IP address _and_ hostname
provides a better audit trail.

Normally this should only take a fraction of a second. However there are
cases where we get a 2(SERVFAIL) after a timeout. If there was just no
PTR record then we'd get a 3(NXDOMAIN) pretty quickly.

Seeing as there is no need to delay the login just to get the hostname,
we can do the IP->hostname lookup in a thread which will do the lookup
and update the utmp table accordingly based on the clients session id.

We move the hostname lookup out of log_login() and into a thread routine
log_utmp_host() which is created from log_login. This allows the login
to complete right away and the utmp table is updated asynchronously.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>