utils: pagination: Add a macro to help with pagination
Add a macro, for_each_table_row(i, pn, datums) to help with creating a
table of data.

This can be used when you have an array or some such that you want to
use as the source for the table data.

e.g

    for_each_table_row(i, pn, datums) {
       lf_set_var(f, "col1", data[i].val1, NULL);
       lf_set_var(f, "col2", data[i].val2, NULL);
       lf_set_var(f, "col3", data[i].val3, de_xss);
       do_zebra(f, i, "listzebra");
       lf_set_row(f, "results");
    }

i      - The loop index
pn     - The pagination structure
datums - The _total_ number of items being displayed across _all_
         pages

i is set to the value of pn.from in the underlying for loop.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>