Define _GNU_SOURCE in individual files
Before, we just set it in common.h which is included in everything else,
however this includes comes after all the system includes and some of
those may need this already defined.

So just define _GNU_SOURCE in each file and remove it from common.h.

This issue was shown up when building under Fedora 20 (GCC 4.9.3) with
numerous compiler warnings/errors.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>