utils: Fix potential issue with adding multipart avars
Under Fedora 20 (GCC 4.8.3) I was seeing the following compiler wanring

    utils.c: In function ‘add_multipart_avar’:
    utils.c:297:22: warning: argument to ‘sizeof’ in ‘snprintf’ call is
     the same expression as the destination; did you mean to provide an
     explicit length? [-Wsizeof-pointer-memaccess]
       snprintf(key, sizeof(key), "%s", token);
                           ^
Key is a char * and thus sizeof will simply return 8. Not what we want.

However we can simply remove the code in question altogether, we don't
need to allocate the variable 'key' which is just has the value of 'token'
copied into it, we can just user token directly.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>