audit: Re-work match_ipv6()
This function is part of the IP address access control list
infrastructure and given an IPv6 address and network/prefixlen will
check whether the address falls within the given network.

This worked fine with standard /64 networks but showed some problems
when dealing with prefix lengths that weren't a multiple of 8.

This commit replaces the tcp wrappers based code with code based on the
ngx_ptocidr() & ngx_http_access_inet6() functions from the nginx web
server.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>