utils: Re-work the hash generation in generate_hash()
Compiling with GCC 9.1.0 gave the following warning

  CC    utils.o
In file included from /usr/include/string.h:638,
                 from utils.c:19:
In function ‘strncat’,
    inlined from ‘generate_hash’ at utils.c:235:3:
/usr/include/bits/string3.h:150:10: warning: ‘__builtin_strncat’ output
may be truncated copying 2 bytes from a string of length 2
[-Wstringop-truncation]
  150 |   return __builtin___strncat_chk (__dest, __src, __len, __bos
(__dest));
      |
^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This is code that copies two bytes of hash string into the full hash
string buffer. I'm not really sure why we get this warning. However we
can replace the calls to sprintf() and strncat() with a single call to
snprintf() to resolve this.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>