common.h: Convert the d_fprintf() macro into an inline function
The only reason this was a macro was so that we could use __func__ to
get the name of the function doing the logging.

This would however be better as a function. So lets just convert it into
a static inline function but then create a macro alias for it that adds
in the __func__ parameter (don't want to have to do that manually each
time we log something). So we still call d_printf() as normal.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>