Re-work config
This largely re-works the configuration of the app.

Rather than just have a bunch of variables *and* #defines for them, put
all the config variables inside a new struct cfg.

We get rid of app_config.h and rename get_config.[ch] -> config.[ch],
config.h contains the struct cfg definition and the default values for
some of the optional config parameters.

We declare struct cfg in app.c and put an extern declaration for it in
common.h as the config is needed by various bits of the code.

This work was prompted by the fact that GCC starting with GCC 10 will
default to enabling -fno-common and while testing this option out I was
presented by many linker errors such as

/usr/bin/ld: audit.o:/home/andrew/programming/fcgi_app_core/src/app_config.h:33: multiple definition of `nr_procs'; get_config.o:/home/andrew/programming/fcgi_app_core/src/app_config.h:33: first defined here

This was due to all the config variables being defined in app_config.h
which was then included in common.h which is included everywhere...

The above work fixes this by having them (in a struct) defined (once) in
app.c and then declared extern in common.h for external access.

This seems a nicer way of doing it, getting rid of the #define variable
aliases.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>