Revert "audit: Fix a segfault when doing name lookups"
This reverts commit ddd96ab7b5ec1bff83bfbf6c18979b0a07fdfd2c.

The above commit was fixing a problem with the log_utmp_host()
threads using the global MYSQL connection handle (via
make_mysql_safe_string().

The way that commit was done was a little convoluted. A much simpler way
is to just declare the connection handle as thread local storage.

So revert the above mentioned commit and then we can do the nicer fix.

Note: This did not revert cleanly, I've tried to keep work done in the
meantime intact...

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>