utils: Use flatePage() in send_template()
This updates send_template() to work with the master branch of
libflate[0] which doesn't know anything about FCGI (it does have the
fmt_function callback for flateSetVar()).

This only requires a minor change to send_template() to have it use
flatePage() to get the processed html template back in a memory buffer
that can then be sent to the client using fcgx_p().

[0]: https://github.com/ac000/libflat

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>